Finish tests for classes, fix bugs
