+++ pydbml/classes/reference.py
@@ -12,10 +12,7 @@ from pydbml.constants import ONE_TO_ONE
 from pydbml.tools import comment_to_dbml
 from pydbml.tools import comment_to_sql
 from pydbml.exceptions import DBMLError
-
-
-if TYPE_CHECKING:
-    from .table import Table
+from pydbml.exceptions import TableNotFoundError
 
 
 class Reference(SQLOjbect):
@@ -24,13 +21,13 @@ class Reference(SQLOjbect):
     It is a separate object, which is not connected to Table or Column objects
     and its `sql` property contains the ALTER TABLE clause.
     '''
-    required_attributes = ('type', 'table1', 'col1', 'table2', 'col2')
+    required_attributes = ('type', 'col1', 'col2')
 
     def __init__(self,
                  type_: Literal[MANY_TO_ONE, ONE_TO_MANY, ONE_TO_ONE],
-                 table1: 'Table',
+                 # table1: 'Table',
                  col1: Union[Column, Collection[Column]],
-                 table2: 'Table',
+                 # table2: 'Table',
                  col2: Union[Column, Collection[Column]],
                  name: Optional[str] = None,
                  comment: Optional[str] = None,
@@ -39,9 +36,9 @@ class Reference(SQLOjbect):
                  inline: bool = False):
         self.schema = None
         self.type = type_
-        self.table1 = table1
+        # self.table1 = table1
         self.col1 = [col1] if isinstance(col1, Column) else list(col1)
-        self.table2 = table2
+        # self.table2 = table2
         self.col2 = [col2] if isinstance(col2, Column) else list(col2)
         self.name = name if name else None
         self.comment = comment
@@ -53,20 +50,17 @@ class Reference(SQLOjbect):
         '''
         >>> c1 = Column('c1', 'int')
         >>> c2 = Column('c2', 'int')
-        >>> t1 = Table('t1')
-        >>> t2 = Table('t2')
-        >>> Reference('>', table1=t1, col1=c1, table2=t2, col2=c2)
-        <Reference '>', 't1'.['c1'], 't2'.['c2']>
+        >>> Reference('>', col1=c1, col2=c2)
+        <Reference '>', ['c1'], ['c2']>
         >>> c12 = Column('c12', 'int')
         >>> c22 = Column('c22', 'int')
-        >>> Reference('<', table1=t1, col1=[c1, c12], table2=t2, col2=(c2, c22))
-        <Reference '<', 't1'.['c1', 'c12'], 't2'.['c2', 'c22']>
+        >>> Reference('<', col1=[c1, c12], col2=(c2, c22))
+        <Reference '<', ['c1', 'c12'], ['c2', 'c22']>
         '''
 
-        components = [f"<Reference {self.type!r}"]
-        components.append(f'{self.table1.name!r}.{[x.name for x in self.col1]!r}')
-        components.append(f'{self.table2.name!r}.{[x.name for x in self.col2]!r}')
-        return ', '.join(components) + '>'
+        col1 = ', '.join(f'{c.name!r}' for c in self.col1)
+        col2 = ', '.join(f'{c.name!r}' for c in self.col2)
+        return f"<Reference {self.type!r}, [{col1}], [{col2}]>"
 
     def __str__(self):
         '''
@@ -82,13 +76,22 @@ class Reference(SQLOjbect):
         Reference(t1[c1, c12] < t2[c2, c22])
         '''
 
-        components = [f"Reference("]
-        components.append(self.table1.name)
-        components.append(f'[{", ".join(c.name for c in self.col1)}]')
-        components.append(f' {self.type} ')
-        components.append(self.table2.name)
-        components.append(f'[{", ".join(c.name for c in self.col2)}]')
-        return ''.join(components) + ')'
+        col1 = ', '.join(f'{c.name!r}' for c in self.col1)
+        col2 = ', '.join(f'{c.name!r}' for c in self.col2)
+        return f"Reference([{col1}] {self.type} [{col2}]"
+
+    def _validate(self):
+        table1 = self.col1[0].table
+        if any(c.table != table1 for c in self.col1):
+            raise DBMLError('Columns in col1 are from different tables')
+        if table1 is None:
+            raise TableNotFoundError('Table on col1 is not set')
+
+        table2 = self.col2[0].table
+        if any(c.table != table2 for c in self.col2):
+            raise DBMLError('Columns in col2 are from different tables')
+        if table2 is None:
+            raise TableNotFoundError('Table on col2 is not set')
 
     @property
     def sql(self):
@@ -99,16 +102,17 @@ class Reference(SQLOjbect):
 
         '''
         self.check_attributes_for_sql()
+        self._validate()
         c = f'CONSTRAINT "{self.name}" ' if self.name else ''
 
         if self.inline:
             if self.type in (MANY_TO_ONE, ONE_TO_ONE):
                 source_col = self.col1
-                ref_table = self.table2
+                ref_table = self.col2[0].table
                 ref_col = self.col2
             else:
                 source_col = self.col2
-                ref_table = self.table1
+                ref_table = self.col1[0].table
                 ref_col = self.col1
 
             cols = '", "'.join(c.name for c in source_col)
@@ -125,14 +129,14 @@ class Reference(SQLOjbect):
             return result
         else:
             if self.type in (MANY_TO_ONE, ONE_TO_ONE):
-                t1 = self.table1
+                t1 = self.col1[0].table
                 c1 = ', '.join(f'"{c.name}"' for c in self.col1)
-                t2 = self.table2
+                t2 = self.col2[0].table
                 c2 = ', '.join(f'"{c.name}"' for c in self.col2)
             else:
-                t1 = self.table2
+                t1 = self.col2[0].table
                 c1 = ', '.join(f'"{c.name}"' for c in self.col2)
-                t2 = self.table1
+                t2 = self.col1[0].table
                 c2 = ', '.join(f'"{c.name}"' for c in self.col1)
 
             result = comment_to_sql(self.comment) if self.comment else ''
@@ -148,11 +152,12 @@ class Reference(SQLOjbect):
 
     @property
     def dbml(self):
+        self._validate()
         if self.inline:
             # settings are ignored for inline ref
             if len(self.col2) > 1:
                 raise DBMLError('Cannot render DBML: composite ref cannot be inline')
-            return f'ref: {self.type} "{self.table2.name}"."{self.col2[0].name}"'
+            return f'ref: {self.type} "{self.col2[0].table.name}"."{self.col2[0].name}"'
         else:
             result = comment_to_dbml(self.comment) if self.comment else ''
             result += 'Ref'
@@ -180,9 +185,9 @@ class Reference(SQLOjbect):
             options_str = f' [{", ".join(options)}]' if options else ''
             result += (
                 ' {\n    '
-                f'"{self.table1.name}".{col1} '
+                f'"{self.col1[0].table.name}".{col1} '
                 f'{self.type} '
-                f'"{self.table2.name}".{col2}'
+                f'"{self.col2[0].table.name}".{col2}'
                 f'{options_str}'
                 '\n}'
             )
